Modified root folder name + refactor
