+++ src/pymodaq_plugins_tango/daq_viewer_plugins/plugins_0D/daq_0Dviewer_Template.py
