+++ src/pymodaq_plugins_tango/daq_viewer_plugins/plugins_1D/__init__.py
