+++ src/pymodaq_plugins_tango/daq_viewer_plugins/plugins_2D/daq_2Dviewer_Template.py
