+++ src/pymodaq_plugins_tango/resources/VERSION
