+++ src/pymodaq_plugins_tango/resources/config_template.toml
